# Conflicts:
#	back-end/src/main/java/com/petmily/backend/community/free/board/FreeBoardController.java
#	back-end/src/main/java/com/petmily/backend/community/free/board/FreeBoardRepository.java
#	back-end/src/main/java/com/petmily/backend/community/free/board/FreeBoardService.java
#	back-end/src/main/java/com/petmily/backend/community/missing/board/MissingBoard.java
#	back-end/src/main/java/com/petmily/backend/community/missing/board/MissingBoardController.java
#	back-end/src/main/java/com/petmily/backend/community/missing/board/MissingBoardDto.java
#	back-end/src/main/java/com/petmily/backend/community/missing/board/MissingBoardRepository.java
#	back-end/src/main/java/com/petmily/backend/community/missing/board/MissingBoardService.java
#	back-end/src/main/java/com/petmily/backend/member/login/service/MemberService.java
#	back-end/src/main/java/com/petmily/backend/support/volunteerReview/repository/VolunteerReviewRepository.java

+++ back-end/src/main/java/com/petmily/backend/comment/domain/Comment.java
@@ -38,12 +38,12 @@ public class Comment {
     @LastModifiedDate
     private LocalDateTime commentUpdate;
 
-    private Integer commentPnum;
+    private Long commentPnum;
 
     private Boolean commentIsSecret;
 
     @Builder
-    public Comment(Member member,String boardId, Long boardNum, String commentContent, Boolean commentIsSecret, Integer commentPnum ){
+    public Comment(Member member,String boardId, Long boardNum, String commentContent, Boolean commentIsSecret, Long commentPnum ){
         this.member = member;
         this.boardId = boardId;
         this.boardNum = boardNum;
